fixed issues with ahts fees on editor
