+++ packages/web/src/pages/VaultEditor/VaultEditorFormPage/steps.ts
@@ -111,7 +111,6 @@ export const AllEditorSections: IEditorSections = {
         component: VaultParametersForm,
         formFields: ["parameters"],
         backButtonTextKey: { creation: "backToVaultDescription" },
-        isAdvanced: true,
       },
       {
         id: "preview",
